64555 - Updating FM36 output model

+++ src/FM36/ESFA.DC.ILR.FundingService.FM36.FundingOutput.Model/Output/LearningDeliveryValues.cs
@@ -50,7 +50,7 @@ namespace ESFA.DC.ILR.FundingService.FM36.FundingOutput.Model.Output
 
         public bool? LearnDel1618AtStart { get; set; }
 
-        public decimal? LearnDelAppAccDaysIL { get; set; }
+        public int? LearnDelAppAccDaysIL { get; set; }
 
         public decimal? LearnDelApplicDisadvAmount { get; set; }
 
@@ -70,9 +70,9 @@ namespace ESFA.DC.ILR.FundingService.FM36.FundingOutput.Model.Output
 
         public bool? LearnDelEligDisadvPayment { get; set; }
 
-        public decimal? LearnDelEmpIdFirstAdditionalPaymentThreshold { get; set; }
+        public int? LearnDelEmpIdFirstAdditionalPaymentThreshold { get; set; }
 
-        public decimal? LearnDelEmpIdSecondAdditionalPaymentThreshold { get; set; }
+        public int? LearnDelEmpIdSecondAdditionalPaymentThreshold { get; set; }
 
         public int? LearnDelHistDaysThisApp { get; set; }
